requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiodns==3.2.0
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==2.1.5
multidict==6.0.5
netifaces==0.10.6
numpy==2.0.0
opencv-python==4.10.0.84
priority==2.0.0
pycares==4.4.0
pycparser==2.22
pydantic==1.10.17
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
Quart==0.19.6
quart-cors==0.7.0
s3transfer==0.10.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
socketio==0.2.1
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.1
watchgod==0.8.2
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
